fixed prototype for size_t fmalloc in acconfig

+++ acconfig.h
@@ -26,11 +26,12 @@ extern "C" {
 #endif
 
 /* Use the clean-failing malloc library in fmalloc.c */
+#include <stddef.h>
 #define xmalloc(s)		fail_die_malloc((s),__FILE__,__LINE__)
 #define xrealloc(p,s)		fail_die_realloc((p),(s),__FILE__,__LINE__)
 #define xfree			free
-void *fail_die_malloc(int, const char *, int);
-void *fail_die_realloc(void *, int, const char *, int);
+void *fail_die_malloc(size_t, const char *, int);
+void *fail_die_realloc(void *, size_t, const char *, int);
 
 /* Prototype strerror() if we don't have it. */
 #ifndef HAVE_STRERROR
